# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRANSPADILLA — Sistema de seguimiento de transporte público en tiempo real
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  correo: string;
  password: string;
}

export type RegisterInputRol = typeof RegisterInputRol[keyof typeof RegisterInputRol];


export const RegisterInputRol = {
  admin: 'admin',
  conductor: 'conductor',
  pasajero: 'pasajero',
} as const;

export interface RegisterInput {
  nombre: string;
  correo: string;
  password: string;
  rol?: RegisterInputRol;
}

export interface Usuario {
  id: number;
  nombre: string;
  correo: string;
  rol: string;
}

export interface AuthResponse {
  token: string;
  usuario: Usuario;
}

export interface RegisterResponse {
  id: number;
  rol: string;
}

export interface Parada {
  id: number;
  nombre: string;
  latitud: number;
  longitud: number;
  /** @nullable */
  orden?: number | null;
}

export interface ParadaInput {
  nombre: string;
  latitud: number;
  longitud: number;
}

export interface AsignarParadaInput {
  parada_id: number;
  orden: number;
}

export interface Ruta {
  id: number;
  nombre: string;
  color: string;
  activa: boolean;
  paradas: Parada[];
}

export interface RutaInput {
  nombre: string;
  color?: string;
}

export interface Bus {
  id: number;
  placa: string;
  estado: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  velocidad?: number | null;
  /** @nullable */
  novedad?: string | null;
  /** @nullable */
  actualizado?: string | null;
  /** @nullable */
  ruta_id?: number | null;
  /** @nullable */
  nombre_ruta?: string | null;
  /** @nullable */
  color_ruta?: string | null;
  /** @nullable */
  conductor_id?: number | null;
  /** @nullable */
  nombre_conductor?: string | null;
}

export interface BusInput {
  placa: string;
  /** @nullable */
  ruta_id?: number | null;
}

export interface GpsInput {
  bus_id: number;
  lat: number;
  lng: number;
  velocidad?: number;
}

export interface NovedadInput {
  bus_id: number;
  novedad: string;
}

export interface FinalizarInput {
  bus_id: number;
}

export interface MessageResponse {
  mensaje: string;
}

export interface Stats {
  totalBuses: number;
  busesActivos: number;
  totalRutas: number;
  totalParadas: number;
  busesConDemora: number;
}