# lib/api-zod/src/generated/types/bus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRANSPADILLA — Sistema de seguimiento de transporte público en tiempo real
 * OpenAPI spec version: 0.1.0
 */

export interface Bus {
  id: number;
  placa: string;
  estado: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  velocidad?: number | null;
  /** @nullable */
  novedad?: string | null;
  /** @nullable */
  actualizado?: string | null;
  /** @nullable */
  ruta_id?: number | null;
  /** @nullable */
  nombre_ruta?: string | null;
  /** @nullable */
  color_ruta?: string | null;
  /** @nullable */
  conductor_id?: number | null;
  /** @nullable */
  nombre_conductor?: string | null;
}