# lib/api-zod/src/generated/types/messageResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRANSPADILLA — Sistema de seguimiento de transporte público en tiempo real
 * OpenAPI spec version: 0.1.0
 */

export interface MessageResponse {
  mensaje: string;
}