# lib/api-zod/src/generated/types/ruta.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRANSPADILLA — Sistema de seguimiento de transporte público en tiempo real
 * OpenAPI spec version: 0.1.0
 */
import type { Parada } from './parada';

export interface Ruta {
  id: number;
  nombre: string;
  color: string;
  activa: boolean;
  paradas: Parada[];
}